conexiones

+++ G4/src/main/java/com/bootcamp/G4/model/MyToken.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class MyToken {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private int id_token;
+    private Long id_token;
     private String name;
     private double price;  
 }
